Implement the dice roll
Implement the roll button to determine which player decides on the play
or the draw.

Add a new view, ALRollView, which handles displaying the result of a
roll. This includes the value rolled and whether the player won.

Abstract the main screen into two ALPlayerView instances. Each instance
holds onto one ALLifeTrackerView and one ALRollView and is responsible
for animating between the two.

Add strong/weak self macros for convenience.